+++ src/serializer.rs
@@ -135,11 +135,7 @@ impl Serializer for Posting {
             write!(writer, " ")?;
         }
 
-        match self.reality {
-            Reality::Real => write!(writer, "{}", self.account)?,
-            Reality::BalancedVirtual => write!(writer, "[{}]", self.account)?,
-            Reality::UnbalancedVirtual => write!(writer, "({})", self.account)?,
-        }
+        self.account.write(writer, settings)?;
 
         if let Some(ref amount) = self.amount {
             write!(writer, "{}", settings.indent)?;
@@ -155,6 +151,21 @@ impl Serializer for Posting {
             for comment in comment.split('\n') {
                 write!(writer, "{}{}; {}", settings.eol, settings.indent, comment)?;
             }
+        }
+
+        Ok(())
+    }
+}
+
+impl Serializer for Account {
+    fn write<W>(&self, writer: &mut W, _settings: &SerializerSettings) -> Result<(), io::Error>
+    where
+        W: io::Write,
+    {
+        match self {
+            Account::Real(name) => write!(writer, "{}", name)?,
+            Account::BalancedVirtual(name) => write!(writer, "[{}]", name)?,
+            Account::UnbalancedVirtual(name) => write!(writer, "({})", name)?,
         }
 
         Ok(())