+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changes
+
+- Merge account name & reality into Account enum
+
 ## [4.1.0] - 2022-01-18
 
 ### Changes